feat-getProductsImpl
* Working on getProduct api client impl.

* Implemented proper api errors.

* GetProductsApi impl ready.

* GetProductApi params updated.

* GetProductsUseCase handled case for no products available.

* Improved code for Jwt token use.

* Optimized DI code.

* Minimum images set to 4 for adding product.

* Added return statements below images requirement in AddProductViewModel.

* Impl sql delight and it's support for GetProductsApi

* Updated ProductRepository to use local product listing as well.

* Impl pagination for getProducts.

* Finalized pagination logic for products in HomeViewModel.

* Added try again logic for loading products on fail.

* Added category id to Product classes.

* Placeholder loading items for products reduced to 2.

+++ src/commonMain/kotlin/onlytrade/app/viewmodel/category/sub/repository/data/db/Subcategory.kt
@@ -1,3 +1,3 @@
 package onlytrade.app.viewmodel.category.sub.repository.data.db
 
-data class Subcategory(val categoryId: Int, val id: Int, val name: String)
+data class Subcategory(val categoryId: Long, val id: Long, val name: String)
